requirements.txt.txt
streamlit
pillow
opencv-python
